# openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Dominion Realm API",
    "version": "0.1.0"
  },
  "paths": {
...
    "/scenes/{scene_id}": {
      "get": {
        "tags": [
          "scenes"
        ],
        "summary": "Scene Detail",
        "operationId": "scene_detail_scenes__scene_id__get",
        "parameters": [
          {
            "name": "scene_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "Scene Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SceneDetail"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "delete": {
        "tags": [
          "scenes"
        ],
        "summary": "Delete Scene",
        "description": "Hard-delete one scene version and everything that points at it. Also purges draft jobs for\nthe same chapter/scene slot. Used by the inbox's bulk 'delete selected'.",
        "operationId": "delete_scene_scenes__scene_id__delete",
        "parameters": [
          {
            "name": "scene_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "Scene Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/DeleteSceneOut"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/scenes/{scene_id}/exemplar": {
      "post": {
        "tags": [
          "scenes"
        ],
        "summary": "Set Exemplar",
        "description": "Mark/unmark this scene as a voice exemplar for its POV (LEARNING_FROM_EDITS Tier 2).\n\nAdds/removes the scene id on the POV's `PovProfile.exemplar_scene_ids` \u2014 the list the drafter\nfew-shots on. Idempotent; disabling on a scene with no profile is a no-op.",
        "operationId": "set_exemplar_scenes__scene_id__exemplar_post",
        "parameters": [
          {
            "name": "scene_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "Scene Id"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ExemplarIn"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "additionalProperties": {
                    "anyOf": [
                      {
                        "type": "string"
                      },
                      {
                        "type": "boolean"
                      }
                    ]
                  },
                  "title": "Response Set Exemplar Scenes  Scene Id  Exemplar Post"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/scenes/{scene_id}/draft-attempts": {
      "get": {
        "tags": [
          "scenes"
        ],
        "summary": "Scene Draft Attempts",
        "description": "Provenance: every preserved stage of this scene's prose pipeline (raw draft, each enrichment\npass, length compress/expand, final rendered), oldest first.",
        "operationId": "scene_draft_attempts_scenes__scene_id__draft_attempts_get",
        "parameters": [
          {
            "name": "scene_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "Scene Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/DraftAttemptOut"
                  },
                  "title": "Response Scene Draft Attempts Scenes  Scene Id  Draft Attempts Get"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/scenes/{scene_id}/versions": {
      "get": {
        "tags": [
          "scenes"
        ],
        "summary": "Scene Versions",
        "description": "Full lineage of a scene: every version sharing its (chapter, scene_no), oldest first.",
        "operationId": "scene_versions_scenes__scene_id__versions_get",
        "parameters": [
          {
            "name": "scene_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "Scene Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/SceneVersionOut"
                  },
                  "title": "Response Scene Versions Scenes  Scene Id  Versions Get"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/scenes/{scene_id}/revert": {
      "post": {
        "tags": [
          "scenes"
        ],
        "summary": "Revert Scene",
        "description": "Roll a scene back to an earlier version: clone that version's prose into a NEW top version\n(versioning is rows, never destructive \u2014 DESIGN \u00a73) and supersede the current one.\n\nThe new version lands APPROVED directly \u2014 reverting is itself the human's decision, so it skips the\ninbox. It deliberately does NOT re-commit the beat's declared deltas (the ledger already reflects\nthe approved state; re-applying relative '+N' deltas would double-count). `scene_id` is the version\nto revert TO.",
        "operationId": "revert_scene_scenes__scene_id__revert_post",
        "parameters": [
          {
            "name": "scene_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "Scene Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SceneOut"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "CritiqueOut": {
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid",
            "title": "Id"
          },
          "reviewer": {
            "type": "string",
            "title": "Reviewer"
          },
          "severity": {
            "type": "string",
            "title": "Severity"
          },
          "note": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Note"
          },
          "payload": {
            "anyOf": [
              {
                "additionalProperties": true,
                "type": "object"
              },
              {
                "type": "null"
              }
            ],
            "title": "Payload"
          },
          "scene_packet_id": {
            "anyOf": [
              {
                "type": "string",
                "format": "uuid"
              },
              {
                "type": "null"
              }
            ],
            "title": "Scene Packet Id"
          }
        },
        "type": "object",
        "required": [
          "id",
          "reviewer",
          "severity"
        ],
        "title": "CritiqueOut"
      },
...
      "DeleteSceneOut": {
        "properties": {
          "deleted": {
            "type": "string",
            "format": "uuid",
            "title": "Deleted"
          },
          "jobs_purged": {
            "type": "integer",
            "title": "Jobs Purged",
            "default": 0
          }
        },
        "type": "object",
        "required": [
          "deleted"
        ],
        "title": "DeleteSceneOut",
        "description": "Result of hard-deleting one scene version."
      },
...
      "DraftAttemptOut": {
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid",
            "title": "Id"
          },
          "stage": {
            "type": "string",
            "title": "Stage"
          },
          "word_count": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Word Count"
          },
          "model": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Model"
          },
          "prose": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Prose"
          },
          "metadata_json": {
            "anyOf": [
              {
                "additionalProperties": true,
                "type": "object"
              },
              {
                "type": "null"
              }
            ],
            "title": "Metadata Json"
          },
          "created_at": {
            "type": "string",
            "format": "date-time",
            "title": "Created At"
          }
        },
        "type": "object",
        "required": [
          "id",
          "stage",
          "created_at"
        ],
        "title": "DraftAttemptOut",
        "description": "One preserved stage of a scene's prose pipeline (provenance: raw \u2192 enrichment \u2192 length \u2192 final)."
      },
...
      "ExemplarIn": {
        "properties": {
          "enabled": {
            "type": "boolean",
            "title": "Enabled"
          }
        },
        "type": "object",
        "required": [
          "enabled"
        ],
        "title": "ExemplarIn",
        "description": "Toggle a scene as a voice exemplar for its POV (LEARNING_FROM_EDITS Tier 2)."
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
      "SceneDetail": {
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid",
            "title": "Id"
          },
          "chapter_id": {
            "type": "string",
            "format": "uuid",
            "title": "Chapter Id"
          },
          "scene_no": {
            "type": "integer",
            "title": "Scene No"
          },
          "version": {
            "type": "integer",
            "title": "Version"
          },
          "status": {
            "type": "string",
            "title": "Status"
          },
          "scene_packet_id": {
            "anyOf": [
              {
                "type": "string",
                "format": "uuid"
              },
              {
                "type": "null"
              }
            ],
            "title": "Scene Packet Id"
          },
          "word_count": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Word Count"
          },
          "length_status": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Length Status"
          },
          "prose": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Prose"
          },
          "prose_source": {
            "type": "string",
            "title": "Prose Source"
          },
          "passes_run": {
            "anyOf": [
              {
                "items": {
                  "type": "string"
                },
                "type": "array"
              },
              {
                "type": "null"
              }
            ],
            "title": "Passes Run"
          },
          "token_count": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Token Count"
          },
          "model": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Model"
          },
          "created_at": {
            "type": "string",
            "format": "date-time",
            "title": "Created At"
          },
          "critiques": {
            "items": {
              "$ref": "#/components/schemas/CritiqueOut"
            },
            "type": "array",
            "title": "Critiques",
            "default": []
          },
          "is_exemplar": {
            "type": "boolean",
            "title": "Is Exemplar",
            "default": false
          }
        },
        "type": "object",
        "required": [
          "id",
          "chapter_id",
          "scene_no",
          "version",
          "status",
          "prose_source",
          "created_at"
        ],
        "title": "SceneDetail"
      },
      "SceneOut": {
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid",
            "title": "Id"
          },
          "chapter_id": {
            "type": "string",
            "format": "uuid",
            "title": "Chapter Id"
          },
          "scene_no": {
            "type": "integer",
            "title": "Scene No"
          },
          "version": {
            "type": "integer",
            "title": "Version"
          },
          "status": {
            "type": "string",
            "title": "Status"
          },
          "scene_packet_id": {
            "anyOf": [
              {
                "type": "string",
                "format": "uuid"
              },
              {
                "type": "null"
              }
            ],
            "title": "Scene Packet Id"
          },
          "word_count": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Word Count"
          },
          "length_status": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Length Status"
          },
          "prose": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Prose"
          },
          "prose_source": {
            "type": "string",
            "title": "Prose Source"
          },
          "passes_run": {
            "anyOf": [
              {
                "items": {
                  "type": "string"
                },
                "type": "array"
              },
              {
                "type": "null"
              }
            ],
            "title": "Passes Run"
          },
          "token_count": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Token Count"
          },
          "model": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Model"
          },
          "created_at": {
            "type": "string",
            "format": "date-time",
            "title": "Created At"
          }
        },
        "type": "object",
        "required": [
          "id",
          "chapter_id",
          "scene_no",
          "version",
          "status",
          "prose_source",
          "created_at"
        ],
        "title": "SceneOut"
      },
...
      "SceneVersionOut": {
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid",
            "title": "Id"
          },
          "chapter_id": {
            "type": "string",
            "format": "uuid",
            "title": "Chapter Id"
          },
          "scene_no": {
            "type": "integer",
            "title": "Scene No"
          },
          "version": {
            "type": "integer",
            "title": "Version"
          },
          "status": {
            "type": "string",
            "title": "Status"
          },
          "scene_packet_id": {
            "anyOf": [
              {
                "type": "string",
                "format": "uuid"
              },
              {
                "type": "null"
              }
            ],
            "title": "Scene Packet Id"
          },
          "word_count": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Word Count"
          },
          "length_status": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Length Status"
          },
          "prose": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Prose"
          },
          "prose_source": {
            "type": "string",
            "title": "Prose Source"
          },
          "passes_run": {
            "anyOf": [
              {
                "items": {
                  "type": "string"
                },
                "type": "array"
              },
              {
                "type": "null"
              }
            ],
            "title": "Passes Run"
          },
          "token_count": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Token Count"
          },
          "model": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Model"
          },
          "created_at": {
            "type": "string",
            "format": "date-time",
            "title": "Created At"
          },
          "agent_original": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Agent Original"
          }
        },
        "type": "object",
        "required": [
          "id",
          "chapter_id",
          "scene_no",
          "version",
          "status",
          "prose_source",
          "created_at"
        ],
        "title": "SceneVersionOut",
        "description": "A scene row plus its preserved pre-edit text, for version diffing in History."
      },
...
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          },
          "input": {
            "title": "Input"
          },
          "ctx": {
            "type": "object",
            "title": "Context"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      }
    }
  }
}